+++ src/WebSocket/Connection.php
@@ -80,14 +80,7 @@ class Connection implements ConnectionInterface
      *
      * @var PingState
      */
-    protected array $pingState = [
-        'last_ping_time' => null,
-        'last_pong_time' => null,
-        'pending_pings' => 0,
-        'ping_count' => 0,
-        'websocket_ping_sent' => false,
-        'pusher_ping_sent' => false,
-    ];
+    protected array $pingState;
 
     /**
      * Event manager instance
@@ -259,16 +252,9 @@ class Connection implements ConnectionInterface
     public function pong(string $type = 'websocket'): void
     {
         $this->pingState['last_pong_time'] = microtime(true);
-
-        if ($this->pingState['pending_pings'] > 0) {
-            $this->pingState['pending_pings']--;
-        }
-
-        if ($type === 'websocket') {
-            $this->pingState['websocket_ping_sent'] = false;
-        } elseif ($type === 'pusher') {
-            $this->pingState['pusher_ping_sent'] = false;
-        }
+        $this->pingState['pending_pings'] = 0;
+        $this->pingState['websocket_ping_sent'] = false;
+        $this->pingState['pusher_ping_sent'] = false;
 
         $this->updateActivity();