+++ tests/bootstrap.php
@@ -135,6 +135,7 @@ $errorHandler = function (int $severity, string $message, string $file, int $lin
     if (!(error_reporting() & $severity)) {
         return false;
     }
+
     throw new ErrorException($message, 0, $severity, $file, $line);
 };
 set_error_handler($errorHandler);